Documentation and removal of unimplemented method
